Refactor graph storage and add persistence tests
